This Version is Almost Clean

+++ src/main/java/com/lemon/spring/interfaces/WebController.java
@@ -2,6 +2,7 @@ package com.lemon.spring.interfaces;
 
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -11,9 +12,9 @@ public interface WebController<K> {
 
     ResponseEntity<Map<String,Object>> update(K entity);
 
-    ResponseEntity<K> findOne(Long id);
+    ResponseEntity<K> findOne(BigInteger id);
 
     ResponseEntity<List<K>> findAll();
 
-    ResponseEntity<Map<String,Object>> delete(Long id);
+    ResponseEntity<Map<String,Object>> delete(BigInteger id);
 }
